refactor(testimonial): singular + api doc
