+++ src/app/testimonial/index.ts
@@ -1,0 +1,2 @@
+export * from './testimonial.component';
+export * from './testimonial.service';
