--- src/app/testimonials/index.ts
@@ -1,1 +1,0 @@
-export * from './testimonials.component';
