+++ src/platform/environment.ts
@@ -7,7 +7,7 @@ import { MessageService } from '../app/message/message.service';
 import { CommitmentService } from '../app/commitment/commitment.service';
 import { MeditationService } from '../app/meditation/meditation.service';
 import { AppointmentService } from '../app/appointment/appointment.service';
-import { TestimonialService } from '../app/testimonials/testimonials.service';
+import { TestimonialService } from '../app/testimonial';
 import { AuthGuard } from '../app/auth-guard';
 import { LoginGuard } from '../app/login-guard';
 import { AdminGuard } from '../app/admin-guard';
